+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+
+## [1.14.0] - 2023-11-02
+
+### Changed
+- Properly set the output dataset in Building Portfolio Recovery Analysis [#423](https://github.com/IN-CORE/pyincore/issues/423)
+- Dependency clean up [#431](https://github.com/IN-CORE/pyincore/issues/431)
+
+### Added
+- Add support for hazard object input from local and remote for building damage analysis [#427](https://github.com/IN-CORE/pyincore/issues/427)
+
+### Fixed
+- CGE warning that using series is deprecated and will raise a type error [#357](https://github.com/IN-CORE/pyincore/issues/357)
+- Pytest fix in workflow [#425](https://github.com/IN-CORE/pyincore/issues/425)
+- Mapping rule to match local repair curve [#438](https://github.com/IN-CORE/pyincore/issues/438)
+- Local tornado x and y axis reversed [#439](https://github.com/IN-CORE/pyincore/issues/439)
+
+
 ## [1.13.0] - 2023-10-11
 
 ### Changed
@@ -14,7 +31,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Add commercial recovery analysis [#395](https://github.com/IN-CORE/pyincore/issues/395)
 - Capability to support for local hazard [#404](https://github.com/IN-CORE/pyincore/issues/404)
 - Add support for local hazard with backward compatibility to analyses [#415](https://github.com/IN-CORE/pyincore/issues/415)
-
 
 ### Fixed
 - Aggregate hazard exposure column for non-structural building damage analysis to avoid column name cutoff and chaining issue with mean damage [#393](https://github.com/IN-CORE/pyincore/issues/393)